Fix for sonar code-smell issues

+++ 2020/day2/src/test/java/org/odoral/adventofcode/y2020/day2/PasswordPhilosophyTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 @Slf4j
 public class PasswordPhilosophyTest {
@@ -21,7 +21,7 @@ public class PasswordPhilosophyTest {
     
     @Test public void test() throws IOException {
         List<String> passwords = passwordPhilosophy.loadPasswords("/scenario1.txt");
-        int counter = passwordPhilosophy.countValidPasswordsForPolicy(passwords, passwordPhilosophy::isValidForFirstPolicy);
+        long counter = passwordPhilosophy.countValidPasswordsForPolicy(passwords, passwordPhilosophy::isValidForFirstPolicy);
         assertEquals(2, counter);
 
         counter = passwordPhilosophy.countValidPasswordsForPolicy(passwords, passwordPhilosophy::isValidForSecondPolicy);
